refactor: move theme selection from profile dropdown to settings page
